Move the LayoutManager Kotlin code to it's own file Fixes #1106
