+++ build-logic/settings.gradle.kts
@@ -24,8 +24,8 @@ dependencyResolutionManagement {
     }
     versionCatalogs { create("libs") { from(files("../gradle/libs.versions.toml")) } }
 }
+
 plugins {
     // Downloads the required Java Toolchain, if needed.
     id("org.gradle.toolchains.foojay-resolver-convention") version "0.8.0"
 }
-
